Abilities defined with explicit subjects
