+++ src/Services/defineAbility.js
@@ -1,6 +1,6 @@
 import { AbilityBuilder, createMongoAbility } from '@casl/ability';
 
-export default function defineAbilityFor(user) {
+function defineAbilityFor(user) {
   const { can, cannot, build } = new AbilityBuilder(createMongoAbility);
   
   if (!user) {
@@ -8,17 +8,49 @@ export default function defineAbilityFor(user) {
   }
 
   const tutorClearance = user.data().clearance;
-  const tutorEmail = user.data().email;
+  // const tutorEmail = user.data().email;
 
   if (tutorClearance === 'admin') {
     can('manage', 'all');
-    cannot('edit', 'Tutor', { email: tutorEmail }).because("Admin cannot change their own clearance");
+    cannot('edit', Tutor, { uid: user.id }).because("Admin cannot change their own clearance");
   }
 
   else if (tutorClearance === "tutor") {
-    can('edit', 'Eval', { owner: user.id });
-    can('edit', 'Assessment', { issued_by: user.id });
+    can('edit', Eval, { owner: user.id });
+    can('edit', Assessment, { issued_by: user.id });
   }
 
-  return build();
+  return build({ detectSubjectType: item => item.constructor });
 }
+
+class Tutor {
+  constructor(dict) {
+    for (const key in dict) {
+      if (dict.hasOwnProperty(key)) {
+        this[key] = dict[key];
+      }
+    }
+  }
+}
+
+class Eval {
+  constructor(dict) {
+    for (const key in dict) {
+      if (dict.hasOwnProperty(key)) {
+        this[key] = dict[key];
+      }
+    }
+  }
+}
+
+class Assessment {
+  constructor(dict) {
+    for (const key in dict) {
+      if (dict.hasOwnProperty(key)) {
+        this[key] = dict[key];
+      }
+    }
+  }
+}
+
+export { defineAbilityFor, Tutor, Eval, Assessment }
